Version 5 Release 🍪
The Version 5  of the Niyoga app has a backend support and all the data get's fetch from a json server. In result, the app will have all the data preserved even if the app is closed.
The custom methods are restored as well. The methods will delete all the data from the database.