v6 Delta 1.5
Deleted unneccessary codes || Update on ERRs - It seems like the errors mentioned in the previous commit is due to some problem in json server and not with this app. Whenever this app is sending put requests or Delete requests repeatedly from a map function, the server disconnects with a err code -98. Otherwise when the amound of tasks are within 15, the app is working completely fine. Hence closing this issue with this commit