Version 6 Release 🧉
In The Version 6  of the Niyoga app the  method's functionality has been changed to marking all task green. On the other side a new method is on board , which will do exactly the opposite of , meaning will remove green from all tasks.

Issue :- Although the app now hot three methods to increase productivity, the backend json server is not really able to handle repeated tens requests. Hence when the app has many tasks listed the methods will not work properly. This will fixed in the next versions, as we will use other servers.